Transitioning NewSheet to be Sheet and consolidating demos
This commit replaces the old sheet, called Sheet by
renaming all references in Cells and Components to
NewSheet as just Sheet.

We also consolidate the cells_demo files to only
create and interact the new Sheet, which is now
the sole kind in the system.
