+++ object_database/web/cells/cells.py
@@ -3061,201 +3061,6 @@ class CodeEditor(Cell):
         self.exportData["keybindings"] = [k for k in self.keybindings.keys()]
 
 
-class NewSheet(Cell):
-    def __init__(
-        self,
-        rowFun,
-        totalColumns,
-        totalRows,
-        colWidth=50,
-        rowHeight=30,
-        numLockRows=0,
-        numLockColumns=0,
-        onCellDblClick=None,
-    ):
-        super().__init__()
-
-        self.rowFun = rowFun
-        self.totalColumns = totalColumns
-        self.totalRows = totalRows
-        self.colWidth = colWidth
-        self.rowHeight = rowHeight
-        if numLockRows >= totalRows:
-            raise "The number of totalRows must be greater than numLockRows."
-        self.numLockRows = numLockRows
-        self.numLockColumns = numLockColumns
-        if numLockColumns >= totalColumns:
-            raise "The number of totalColumns must be greater than numLockColumns."
-
-        # TODO: Add double click feature from
-        # current old sheet
-
-    def recalculate(self):
-        self.exportData["numLockRows"] = self.numLockRows
-        self.exportData["numLockColumns"] = self.numLockColumns
-        self.exportData["totalColumns"] = self.totalColumns
-        self.exportData["totalRows"] = self.totalRows
-        self.exportData["colWidth"] = self.colWidth
-        self.exportData["rowHeight"] = self.rowHeight
-
-    def onMessage(self, msgFrame):
-        if msgFrame["event"] == "sheet_needs_data":
-            requested_frames = msgFrame["frames"]
-            response_frames = []
-            for frame in requested_frames:
-                rows_to_send = self.rowFun(
-                    # start_row
-                    frame["origin"]["y"],
-                    # end row
-                    frame["corner"]["y"],
-                    # start_column
-                    frame["origin"]["x"],
-                    # end_column
-                    frame["corner"]["x"],
-                )
-                response_frames.append(
-                    {
-                        "data": rows_to_send,
-                        "origin": frame["origin"],
-                        "corner": frame["corner"],
-                    }
-                )
-
-            dataToSend = [{"action": msgFrame["action"], "frames": response_frames}]
-            if self.exportData.get("dataInfo") is None:
-                self.exportData["dataInfo"] = dataToSend
-            else:
-                self.exportData["dataInfo"] += dataToSend
-            self.wasDataUpdated = True
-            self.markDirty()
-
-
-class Sheet(Cell):
-    """A spreadsheet viewer. The dataset must be static."""
-
-    def __init__(
-        self,
-        rowFun,
-        totalColumns,
-        totalRows,
-        colWidth=50,
-        rowHeight=30,
-        numLockRows=0,
-        numLockColumns=0,
-        onCellDblClick=None,
-    ):
-        """
-        rowFun: func
-            function taking integer 'start_row' and 'end_row' row indexes and
-            'start_column' and 'end_column' that
-            returns a list of rows, themselves list of values, to populate the
-            table; Note the first value, i.e. row[0], of each row is the named index
-        totalColumns: int
-            total number of columns
-        totalRows: int
-            total number of rows
-        colWidth: int
-            height of columns in pixels
-        rowHeight: int
-            height of row in pixels
-        onCellDblClick: str
-            function to run after user double clicks a cell. It takes as keyword
-            arguments row, col, and sheet where row and col represent the row and
-            column clicked and sheet is the Sheet object. Clicks on row(col)
-            headers will return row(col) values of -1
-        """
-        super().__init__()
-
-        self.rowFun = rowFun
-        self.totalColumns = totalColumns
-        self.totalRows = totalRows
-        self.colWidth = colWidth
-        self.rowHeight = rowHeight
-        if numLockRows >= totalRows:
-            raise "The number of totalRows must be greater than numLockRows."
-        self.numLockRows = numLockRows
-        self.numLockColumns = numLockColumns
-        if numLockColumns >= totalColumns:
-            raise "The number of totalColumns must be greater than numLockColumns."
-        self.dataInfosToSend = []
-
-        self.error = Slot(None)
-        self._overflow = "auto"
-
-        self._hookfns = {}
-        if onCellDblClick is not None:
-
-            def _makeOnCellDblClick(func):
-                def _onMessage(sheet, msgFrame):
-                    return onCellDblClick(
-                        sheet=sheet, row=msgFrame["row"], col=msgFrame["col"]
-                    )
-
-                return _onMessage
-
-            self._hookfns["onCellDblClick"] = _makeOnCellDblClick(onCellDblClick)
-
-    def _addHandsontableOnCellDblClick(self):
-        pass
-
-    def recalculate(self):
-        errorCell = Subscribed(
-            lambda: Traceback(self.error.get()) if self.error.get() is not None else Text("")
-        )
-        self.children["error"] = errorCell
-
-        # Deleted the postscript that was here.
-        # Should now be implemented completely
-        # in the JS side component.
-
-        self.exportData["colWidth"] = self.colWidth
-        self.exportData["rowHeight"] = self.rowHeight
-        self.exportData["totalColumns"] = self.totalColumns
-        self.exportData["totalRows"] = self.totalRows
-        self.exportData["numLockRows"] = self.numLockRows
-        self.exportData["numLockColumns"] = self.numLockColumns
-        self.dataInfosToSend = []
-
-        # self.exportData['handlesDoubleClick'] = ("onCellDblClick" in self._hookfns)
-
-    def onMessage(self, msgFrame):
-        """TODO: We will need to update the Cell lifecycle
-        and data handling before we can move this
-        to the JS side"""
-
-        # Note: this used to be
-        # 1, using the old assumption
-        # that rows would not come over with
-        # their first column values set.
-        # ROW_LEN_OFFSET = 0
-
-        if msgFrame["event"] == "sheet_needs_data":
-            frame = msgFrame["frame"]
-            start_row = frame["origin"]["y"]
-            end_row = frame["corner"]["y"]
-            start_column = frame["origin"]["x"]
-            end_column = frame["corner"]["x"]
-            rowsToSend = self.rowFun(start_row, end_row, start_column, end_column)
-            dataInfo = {
-                "data": rowsToSend,
-                "action": msgFrame["action"],
-                "origin": frame["origin"],
-                "corner": frame["corner"],
-                "name": frame.get("name"),
-                "response_id": msgFrame["request_index"],
-            }
-            # stage this piece of data to be sent when we recalculate
-            if self.exportData.get("dataInfo") is None:
-                self.exportData["dataInfo"] = [dataInfo]
-            else:
-                self.exportData["dataInfo"].append(dataInfo)
-
-            self.wasDataUpdated = True
-            self.markDirty()
-        else:
-            return self._hookfns[msgFrame["event"]](self, msgFrame)
-
-
 class Plot(Cell):
     """Produce some reactive line plots."""
